backend/api/handlers.go
// backend/api/handlers.go

package api

import (
	"github.com/gin-gonic/gin"
	"net/http"
)

func HandleUpload(c *gin.Context) {
	// Placeholder URL (replace with actual logic)
	url := "https://example.com/uploads/image123.jpg"

	// Respond with the URL
	c.JSON(http.StatusOK, gin.H{
		"message": "Image uploaded successfully!",
		"url":     url,
	})
}

func HandleImage(c *gin.Context) {
	// Implement image retrieval logic here
	// Retrieve image information from the database
	// Serve the image or thumbnail based on the request
}
